Listhias/requirements.txt
asgiref-3.8.1
django-5.1.4
sqlparse-0.5.3

djangorestframework-simplejwt-5.3.1
pyjwt-2.10.1

